Initial implementation of BalanceManager
- integrate Eosio SDK
- fetch all balances from SDK and save them to local database
- adopt demo app GUI for displaying account balances

+++ EosKitDemo/EosKitDemo/Adapters/IAdapter.swift
@@ -8,7 +8,7 @@ protocol IAdapter {
 
     var lastBlockHeight: Int? { get }
     var syncState: EosKit.SyncState { get }
-    var balance: Double { get }
+    var balance: Decimal { get }
 
     var receiveAddress: String { get }
